dev-requirements.txt
pyinstaller
pyqt6rc
black
